examples/python/web3py-scripts/requirements.txt
web3==6.11.3
py-solc-x==2.0.2
eth-account==0.10.0
eth-utils==2.3.1
hexbytes==0.3.1
